chore: move wallet-related files to dedicated folder (#6)
