--- src/auth/dto/disconnect-wallet.dto.ts
@@ -1,3 +1,0 @@
-export class DisconnectWalletDto {
-  address?: string;
-};
