--- src/auth/dto/wallet-status.dto.ts
@@ -1,4 +1,0 @@
-export class walletStatusDto {
-  connected: boolean;
-  addresses: string;
-}
