+++ src/wallet/wallet.module.ts
@@ -1,0 +1,13 @@
+import { Module } from '@nestjs/common';
+import { PrismaModule } from 'nestjs-prisma';
+import { WalletController } from './wallet.controller';
+import { WalletService } from './wallet.service';
+
+@Module({
+  imports: [
+    PrismaModule,
+  ],
+  controllers: [WalletController],
+  providers: [WalletService]
+})
+export class WalletModule { }
